filestate/upgrade: Support backfilling projects for detached stack
Adds an option to the filestate upgrade operation that, when supplied,
allows the caller to fill project names for stacks
for which we could not guess a project name.

The caller supplies a function with the following signature,
taking a list of stack names and returning a list of project names
in the same order.

    func(stacks []Name) (projects []Name, err error)

A caller like the CLI can use this hook to prompt the user for input,
allowing users to fill in project names for stacks.