test: Add failing test case for #12593
Information in the original issue is a bit limited,
but we know that the error occured when the user
passed a `provider=x` to a multi-language component,
which then instantiate a resource handled by that provider
with itself as the parent.
The expected behavior was for the resource to use the explicit provider,
but instead, it used the default global provider.

The reproduction does roughly the same for Go, Python, and Node,
except it instantiates three resources:

1. using the default provider
2. using an explicit provider specified with `provider=x`
3. using an explicit provider specified with `providers=[x]`
3. using an explicit provider specified with `providers={pkg: x}`

Case (2) fails as expected from #12593 for Python.

```
--- PASS: TestConstructProviderPropagationGo (14.66s)
--- PASS: TestConstructProviderPropagationNode (15.03s)
--- FAIL: TestConstructProviderPropagationPython (17.63s)
```