sdk/py/Resource: Don't drop merged providers
We previously fixed an issue where a provider specified on a component
resource with `provider=` was not correctly propagated to its children
in the providers bag. (#12161)
The fix was based on setting `resource._provider` only if
the resource package matched the provider package.
Otherwise, the provider would be added to the `providers` bag
to be propagated to the children.

This introduced a bug where we would drop the provider somehow,
and start using the default provider instead.
The change was reverted. (#12520)

The reason for this failure was that the resulting providers bag
was being ignored:

- The bag returned by `_get_providers` is stored on the resulting
  resource:

  https://github.com/pulumi/pulumi/blob/f76bd5463b512222b62d9e60191d93ab2a66290e/sdk/python/lib/pulumi/resource.py#L866

- A request is made to `register_resource`, which calls
  `prepare_resource`:

  https://github.com/pulumi/pulumi/blob/f76bd5463b512222b62d9e60191d93ab2a66290e/sdk/python/lib/pulumi/resource.py#L880
  https://github.com/pulumi/pulumi/blob/f76bd5463b512222b62d9e60191d93ab2a66290e/sdk/python/lib/pulumi/runtime/resource.py#L811

- `register_resource` re-interprets the `provider` and `providers`
  resource options, based on what's in the original `opts` struct

  https://github.com/pulumi/pulumi/blob/f76bd5463b512222b62d9e60191d93ab2a66290e/sdk/python/lib/pulumi/runtime/resource.py#L210

To fix this, we modify the `opts` struct after merging the providers
in `_get_providers`.