doc(sdk/go): Clarify purpose of ResourceOptions, InvokeOptions
A user expressed confusion about the existence of ResourceOptions and
ResourceOption in Community Slack, because it wasn't clear which to use.

Clarify in the documentation that these are not intended to be passed
to a resource constructor or a provider function,
and are merely snapshots of what you do pass into them.