Fix config set-all saving secret provider configuration
Clean up some more hidden loading of config files in
getStackEncrypter/getStackDecrypter that could cause the config file to
change due to secret provider initialization. Importantly this was
missed in config set-all where we would load the config file, then later
on call getStackEncrypter which would initialize the secrets config and
save it to disk, but not return that information back up to "set-all"
which then overwrote it with the config structure that it had been
building. This could result in secret values written to the config file
but without saving their salt meaning they could never correctly be
decrypted.

Fixes #12632.