[sdkgen/py] Fix referencing types with diff pkg name
The Python SDK generator wasn't correctly referencing types that have a
different package name in their type token than the name of the schema's
package (which is the case when embedding k8s CRD resources/types in a
package). This change fixes the generator to emit these local references
correctly.