sdk/go: Don't drop Dependencies in ResourceOptions snapshot for MLCs
We recently added a new API, NewResourceOptions to build a preview
of a set of resource options.

    func NewResourceOptions(...ResourceOption) (*ResourceOptions, error)

This currently drops the dependencies of an MLC from the snapshot.
The reason for this is that it used a special type "urnSet"
to represent dependencies for MLCs received over the wire.
This decision was made at the time because the original Resource objects
were not available for these dependencies.

Turns out that that limitation isn't a blocker:
we can use newDependencyResource to create dummy Resource objects
that hold nothing but a URN.

This allows NewResourceOptions to work on options even inside an MLC.
Note that this currently only works for some of the options:
those that are propagated from `construct` into the Go SDK options.
Others will be added as part of #12154.

Testing:
The accompanying test failed for Dependencies without this change.