test(go/host): GetRequiredPlugins and GetProgramDependencies
Adds a test that sets up a Go module with a few dependencies
and queries its plugins and dependencies with the Go language host.
This test passes today because we're running in module mode.

A sibling test will be added for vendor mode
to verify the fix for #12526, where this same check would fail
if we vendor the dependencies into the Pulumi program.

NOTE:
The test does not run in parallel because we need to change the
working directory. This is resolved in a follow-up.