Support SSH-style Git URLs
This adds support for SSH-style Git URLs, enabling folks to use
private repos for their templates.

For instance,

    $ pulumi new git@github.com:acmecorp/templates/website

will now work as intended.

The logic to find the SSH key depends on the well-known ~/.ssh/config
file when available (as that is typically how the ssh-agent is
configured for auth with the Git client), but falls back on other
common key names in the event that isn't configured.

If the SSH key is protected by a password, the user will be prompted
to supply the password on-demand. (It is memoized to avoid asking
multiple times, as the template workflow requires using it more than
once.) To avoid prompting, the PULUMI_GITSSH_PASSPHRASE envvar can
be set -- which is coincidentally the only option in non-interactive
modes.

Fixes #4872 and #5007.