Fix race condition in go/generate function.
We were building pulumi and hit an error in this function when running the makefile, but only in certain environments.
I read the code, and it looks like there's a race condition here where the gofmt command finishes
before the io.Copy starts, giving us this error:

⚠️  x86_64    | + make build
ℹ️  x86_64    | BUILD:
ℹ️  x86_64    | go generate ./pulumi/...
⚠️  x86_64    | panic: unexpected error running gofmt: read |0: file already closed

The go async function is unnecessary here, and the Go exec docs explain that StderrPipe should not be used
with Run: https://pkg.go.dev/os/exec#Cmd.StderrPipe

CombinedOutput is a simpler way to accomplish the same thing, and fixes the race condition.

Signed-off-by: Dan Lorenc <dlorenc@chainguard.dev>