[sdk/nodejs] Fix Parent/NoParent aliases
The Node.js SDK is not sending the right `Parent` and `NoParent` alias
spec.

https://github.com/pulumi/pulumi/blob/117955ce14b6cf2dd093f8356ec7e26427aca4bf/sdk/nodejs/runtime/resource.ts#L332-L361

- It does not set `NoParent` when it should, but instead sets `Parent`
  to empty.

- It sets `NoParent` to true and leaves `Parent` empty when both the
  alias and resource have no `Parent` specified.

- If the resource has a parent and `Parent` isn't set on the alias, it
  sends the parent's URN for the `Parent`.

This commit fixes these issues so that the Node.js SDK sends correctly
specified alias specs.

However, we continue to set the alias's `Parent` to the resource's
parent if the alias does not specify a `Parent`. We shouldn't have to do
this, but doing it allows us to maintain backwards compatibility in the
case when the Node.js SDK is upgraded, but not the CLI.