CONTRIBUTING: Clarify gofumpt further
We mention that you need to install gofumpt,
but it's easy to miss that there's some editor setup involved.

gofumpt is not mentioned in the `brew install` which,
depending on editor, may be necessary.
(Editors that use gopls don't need to install it globally,
but editors that invoke it manually need it installed.)

Lastly, Submitting a PR could do with explaining the lint/gofumpt
requirements.

Ref: https://github.com/pulumi/pulumi/pull/12854#discussion_r1190467688