pkg/providers: Extract plugin installation to pkg/workspace
This adds a new pkg/workspace package
and moves the logic to install a provider into a function
in this new package.

This renames that functionality and the accompanying error to
InstallPlugin, as we intend to reuse the same logic to download other
plugins.

This commit contains no behavioral changes.