Remove Useless Constructors.
This commit enables a lint that identifies constructions which have
no effect. Namely, constructors with no body, or constructors that
simply call super(). Removing the constructor is equivalent to having
an empty constructor, and thus the code is unnecessarily complex.
There is only one instance of our codebase of a useless constructor.

See https://docs.rome.tools/lint/rules/nouselessconstructor/ for more.