fix(sdkgen/go): Don't set default to zero if env is empty
sdkgen has a discrepancy in behavior for default values.
There are three scenarios:

- default value
- default value and an environment variable fallback
- only an environment variable fallback

We handle the first two correctly:
the value is set only if there was something to set it to.
But in the third case, we set variables to their zero value
when the environment variable is empty.

    // Current implementation effectively does:

    if v := os.Getenv("whatever"); v != "" {
        out.Field = pointerOf(v)
    } else {
        out.Field = pointerOf("")
    }

To fix this, we generate code to set these values
only if the environment variable is non-empty.

Implementation notes:
Previosly, getDefaultValue generated an expression
holding the default value.
For fixed default values, it was just the value.
For default values with environment variable fallback,
it took the form:

    getEnvOrDefault($default, $parser, $env1, $env2, ...).($type)

For example:

    getEnvOrDefault("", nil, "ITEM_NAME").(string)

Given that getDefaultValue returned an expression,
its consumers were able to do something like:

    if result == nil {
        result = getEnvOrDefault(...).(string)
    }

As part of this fix, we need to stop assinging to result
if the environment variable is unset,
so we want getDefaultValue to generate:

    if result == nil {
        if d := getEnvOrDefault(...); d != nil {
            result = d
        }
    }

To accomplish this, we switch getDefaultValue to setDefaultValue.
Instead of returning an expression, it accepts a callback.
It generates the blocks it needs to, and invokes the callback
with an expression holding the default value if necessary.

Resolves #12971