Revert change in targeted update behavior.
Fixes #12964

This commit uses an untargeted resource's old inputs directly and does
not send the untargeted resource's inputs to the provider for
alterations in `provider.Check()`.

This commit fixes a bug in behavior where `provider.Check()` was
no longer be called on untargeted resources. `provider.Check()` would
alter the engine's inputs for use in resource.Goals and skipping this
would cause `__defaults` to not be configured properly in same steps by
`provider.Check()` and cause replaces in some cases.