ci: make lint_golang to lint all Go packages
Adds a `make lint_golang` target that lints all Go packages
in one command instead of copy-pasting between them.

Instead of duplicating the list of packages in ci-lint,
use the golangci-lint-action to install the linter once,
and let the `make` target actually run it.

The primary motivation for this is to reduce duplication in CI setup
in anticipation of adding a new package to the list.

Usage notes:

- The `--path-prefix` flag ensures that file paths
  in the error messages are correct.
- The optional GOLANGCI_LINT_ARGS variable allows adding additional
  arguments.
  We use this to opt-into github-actions-friendly output in CI.