[logging] Improve filter performance
These changes improve the throughput and reduce the allocation volume of
the filters used to redact config and credential secrets from logs and
property values by replacing the use of strings.Replacer with an
implementation based on the Aho-Corasick machine. In the common case
where a value does not contain any secrets, the former allocates a new
copy of the string while the former does not.

In several real-world tests, this reduces the CLI's overall allocation
volume by about 25%.