test(sdk/npm): Don't download from NPM or Yarn regstries
Tests that download from NPM or Yarn will be brittle, unsafe, and slow.
We don't need to hit these servers directly.

This change alters the package installation tests to
set up a local HTTP server that implements a subset of the
NPM Package Registry API [1] -- just enough to satisfy these tests.

  [1]: https://github.com/npm/registry/blob/master/docs/REGISTRY-API.md

In the tests, we'll set `$HOME` to a new temporary directory and write
an `~/.npmrc` or `~/.yarnrc` requesting use of the local server as the
registry.

With this, the tests are entirely local. No external requests.