Fix inconsistency in calling DiffConfig
Diff/DiffConfig is currently always called with the old outputs and new inputs,
except for this one case in step_generator where we called DiffConfig
with the old inputs instead of old outputs.

For providers inputs and outputs are always the same (Configure doesn't
allow providers to return any properties, so the engine enforces that
outputs=inputs) so using inputs or outputs here doesn't have any effect,
but less confusing to not have this inconsistency.